requirements.txt
discord.py>=2.4.0,<3
aiohttp>=3.9.5
python-dotenv>=1.0.1
